Final changes from bringup. Tests passing

+++ runtest.sh
@@ -23,38 +23,54 @@ rm -f $DIR/simtrace
 mkfifo $DIR/simtrace
 #timeout -s9 $TIMEOUT $DIR/$MODEL/build/top +dramcfg=$DRAMCFG +memfile=$HEXFILE +tracefile=simtrace +uartfile=$UARTFILE +logfile=$LOGFILE &
 $DIR/rtl/build/top \
-    --testplusarg dramcfg=$DRAMCFG \
-    --testplusarg memfile=$HEXFILE \
-    --testplusarg tracefile=$DIR/simtrace \
-    --testplusarg uartfile=$UARTFILE \
-    --testplusarg logfile=$LOGFILE &
+    +dramcfg=$DRAMCFG \
+    +memfile=$HEXFILE \
+    +tracefile=$DIR/simtrace \
+    +uartfile=$UARTFILE \
+    +logfile=$LOGFILE &
 SIMPID=$!
 
 $DIR/runspike.sh --log-commits --cosim=$DIR/simtrace $ELFFILE 2>/dev/null &
 SPIKEPID=$!
 
-ERROR=0
+function fail {
+    echo "TEST FAILED"
+    rm -f $DIR/simtrace
+    exit 1
+}
 
-wait $SIMPID
-if [ $? -ne 0 ]; then
-    echo "ERROR: rtl returned non-zero"
-    ERROR=1
+SIMSTATUS=
+SPIKESTATUS=
+until [ -n "$SIMSTATUS" ] && [ -n "$SPIKESTATUS" ]; do
+    wait -n -p PID $SIMPID $SPIKEPID
+    STATUS=$?
+    case $PID in
+        $SIMPID) SIMSTATUS=$STATUS; SIMPID=;;
+        $SPIKEPID) SPIKESTATUS=$STATUS; SPIKEPID=;;
+        *) echo "ERROR: Unexpected return value from wait: $PID"; fail;;
+    esac
+    if [ $STATUS -ne 0 ]; then break; fi
+done
+
+if [ -n "$SIMSTATUS" ] && [ $SIMSTATUS -ne 0 ]; then
+    echo "ERROR: rtl returned non-zero status: $SIMSTATUS"
+    fail
 fi
 
-wait $SPIKEPID; SPIKESTATUS=$?
-if [ $SPIKESTATUS -ne 0 ]; then
-    echo "ERROR: spike exited with non-zero status"
-    ERROR=1
+if [ -n "$SPIKESTATUS" ] && [ $SPIKESTATUS -ne 0 ]; then
+    echo "ERROR: spike exited with non-zero status: $SPIKESTATUS"
+    fail
+fi
+
+if [ -z "$SIMSTATUS" ] || [ -z "$SPIKESTATUS" ]; then
+    echo "ERROR: missing return status from sim or spike"
+    fail
+fi
+
+$DIR/checkmem.py $LOGFILE
+if [ $? -ne 0 ]; then
+    fail
 fi
 
 rm -f $DIR/simtrace
-$DIR/checkmem.py $LOGFILE
-if [ $? -ne 0 ]; then
-    ERROR=1
-fi
-
-if [ $ERROR -ne 0 ]; then
-    echo "TEST FAILED"
-fi
-
-exit $ERROR
+exit 0
